+++ src/server/webpack-config.js
@@ -3,17 +3,36 @@ import fsPath from 'path';
 import { rootModulePath } from './paths';
 
 const NODE_MODULES_PATH = fsPath.join(rootModulePath(), 'node_modules');
+const UIHARNESS_ENTRY = fsPath.join(__dirname, '../client/entry');
 
 
-const babelLoader = (extension) => ({
-  // See: https://github.com/babel/babel-loader#options
-  loader: 'babel',
-  test: extension,
-  exclude: /(node_modules|bower_components)/,
-  query: {
-    // plugins: ['./build/babelRelayPlugin'],
-  },
-});
+
+// HACK (Relay).
+//       Prevent error with `fetch` which attempts to look for a `self` object.
+//       This occurs when parsing the `react-relay` module on the server while compiling.
+global.self = { fetch: null };
+
+
+
+const babelLoader = (extension, isRelayEnabled) => {
+  const loader = {
+    // See: https://github.com/babel/babel-loader#options
+    loader: 'babel',
+    test: extension,
+    exclude: /(node_modules|bower_components)/,
+    query: {
+      plugins: [],
+    },
+  };
+
+  // Add optional plugins.
+  if (isRelayEnabled) {
+    loader.query.plugins.push(fsPath.resolve('./src/relay/babel-relay-plugin'));
+  }
+
+  // Finish up.
+  return loader;
+};
 
 
 
@@ -21,22 +40,33 @@ const babelLoader = (extension) => ({
  * Creates a Webpack compiler instance.
  *
  * @param {Object} options:
- *            - entry:          Array of entry paths.
- *            - isProduction:   Flag indicating if the builder is in production mode.
- *                              Default false.
+ *            -- entry:           Array of entry paths.
+ *            -- isProduction:    Flag indicating if the builder is in production mode.
+ *                                Default: false.
+ *
+ *            -- outputFile:      The name of the output file.
+ *                                Default: 'bundle.js'.
+ *
+ *            -- isRelayEnabled:  Flag indicating if relay is being used.
+ *                                Default: false.
  *
  * @return {Object} compiler.
  */
 export default (options = {}) => {
   const isProduction = options.isProduction || false;
+  const outputFile = options.outputFile || 'bundle.js';
+  const isRelayEnabled = options.isRelayEnabled || false;
 
   const config = {
-    entry: options.entry,
-    output: { path: '/', filename: 'bundle.js' },
+    entry: {
+      app: options.entry,
+      vendor: ['react', 'react-dom', 'react-relay', UIHARNESS_ENTRY],
+    },
+    output: { path: '/', filename: outputFile },
     module: {
       loaders: [
-        babelLoader(/\.js$/),
-        babelLoader(/\.jsx$/),
+        babelLoader(/\.js$/, isRelayEnabled),
+        babelLoader(/\.jsx$/, isRelayEnabled),
         { test: /\.json$/, loader: 'json-loader' },
         { test: /\.(png|svg)$/, loader: 'url-loader' },
       ],
@@ -48,12 +78,17 @@ export default (options = {}) => {
       resolveLoader: { fallback: NODE_MODULES_PATH },
     },
     plugins: [
-      // https://github.com/webpack/docs/wiki/optimization#deduplication
+      // Remove duplicate code.
+      //    See - https://github.com/webpack/docs/wiki/optimization#deduplication
       new webpack.optimize.DedupePlugin(),
 
       // [Moment.js] only load subset of locales to reduce size.
-      //   See - http://stackoverflow.com/a/25426019/1745661
+      //    See - http://stackoverflow.com/a/25426019/1745661
       new webpack.ContextReplacementPlugin(/moment[\/\\]locale$/, /en/),
+
+      // Break out common libs into their own code-chunk.
+      //    See - https://webpack.github.io/docs/code-splitting.html#split-app-and-vendor-code
+      new webpack.optimize.CommonsChunkPlugin({ name: 'vendor', filename: 'vendor.js' }),
     ],
   };