+++ start.js
@@ -1,10 +1,12 @@
-"use strict"
-var R = require("ramda");
-var minimist = require("minimist");
-var chalk = require("chalk");
-var server = require("./lib/server");
+'use strict'
+require('babel-polyfill');
+var R = require('ramda');
+var minimist = require('minimist');
+var chalk = require('chalk');
+var server = require('./lib/server');
 var log = require('./lib/shared/log').default;
 
+// Read-in command-line arguments.
 var args = process.argv.slice(2);
 args = args.length > 0 ? args = minimist(args) : {};
 
@@ -25,14 +27,14 @@ args = args.length > 0 ? args = minimist(args) : {};
  */
 if (R.is(String, args.entry)) {
   server.start({
-    entry: args.entry.split(","),
+    entry: args.entry.split(','),
     port: args.port
   })
   .catch(err => {
-    log.error(chalk.red("Failed to start."));
+    log.error(chalk.red('Failed to start.'));
     log.error(chalk.red(err.message));
     log.error()
   });
 } else {
-  log.error(chalk.red("No entry path was specified, for example: `--entry ./src/specs`\n"));
+  log.error(chalk.red('No entry path was specified, for example: `--entry ./src/specs`\n'));
 }