+++ test/shared/ThisContext-props.test.js
@@ -1,11 +1,11 @@
-import _ from "lodash";
-import { expect } from "chai";
-import ThisContext from "../../src/shared/ThisContext";
-import bdd from "../../src/shared/bdd";
-import api from "../../src/shared/api-internal";
+import _ from 'lodash';
+import { expect } from 'chai';
+import ThisContext from '../../src/shared/ThisContext';
+import bdd from '../../src/shared/bdd';
+import api from '../../src/shared/api-internal';
 
 
-describe("ThisContext", () => {
+describe('ThisContext', () => {
   let suite, context;
   afterEach(() => { bdd.reset(); })
   beforeEach(() => {
@@ -16,29 +16,29 @@ describe("ThisContext", () => {
   });
 
 
-  describe("cropMarks()", () => {
-    it("has crop-marks by default", () => {
+  describe('cropMarks()', () => {
+    it('has crop-marks by default', () => {
       expect(context.cropMarks()).to.equal(true);
       expect(context.cropMarks.size()).to.equal(25);
       expect(context.cropMarks.offset()).to.equal(5);
     });
 
-    it("stores crop-marks value", () => {
+    it('stores crop-marks value', () => {
       context.cropMarks(true).cropMarks(false);
       expect(context.cropMarks()).to.equal(false);
     });
 
-    it("persists to the [current] map", () => {
+    it('persists to the [current] map', () => {
       context
         .cropMarks(true)
         .cropMarks.size(50)
         .cropMarks.offset(0);
-      expect(api.current.get("cropMarks")).to.equal(true);
-      expect(api.current.get("cropMarks.size")).to.equal(50);
-      expect(api.current.get("cropMarks.offset")).to.equal(0);
+      expect(api.current.get('cropMarks')).to.equal(true);
+      expect(api.current.get('cropMarks.size')).to.equal(50);
+      expect(api.current.get('cropMarks.offset')).to.equal(0);
     });
 
-    it("stores extensions (size/offset)", () => {
+    it('stores extensions (size/offset)', () => {
       context
         .cropMarks.size(50)
         .cropMarks.offset(0);
@@ -46,7 +46,7 @@ describe("ThisContext", () => {
       expect(context.cropMarks.offset()).to.equal(0);
     });
 
-    it("throws on invalid values", () => {
+    it('throws on invalid values', () => {
       expect(() => { context.cropMarks(123) }).to.throw();
       expect(() => { context.cropMarks.size(true) }).to.throw();
       expect(() => { context.cropMarks.offset(true) }).to.throw();
@@ -54,13 +54,13 @@ describe("ThisContext", () => {
   });
 
 
-  describe("size (width/height)", function() {
-    it("has not width/height by default ('auto')", () => {
-      expect(context.width()).to.equal("auto");
-      expect(context.height()).to.equal("auto");
+  describe('size (width/height)', function() {
+    it('has not width/height by default ("auto")', () => {
+      expect(context.width()).to.equal('auto');
+      expect(context.height()).to.equal('auto');
     });
 
-    it("stores width/height (number)", () => {
+    it('stores width/height (number)', () => {
       context
         .width(250)
         .height(120)
@@ -68,28 +68,28 @@ describe("ThisContext", () => {
       expect(context.height()).to.equal(120);
     });
 
-    it("resets with `null`", () => {
+    it('resets with `null`', () => {
       context.width(250).height(120)
       context.width(null).height(null)
-      expect(context.width()).to.equal("auto");
-      expect(context.height()).to.equal("auto");
+      expect(context.width()).to.equal('auto');
+      expect(context.height()).to.equal('auto');
     });
 
-    it("width throws if a number of string is not passed", () => {
+    it('width throws if a number of string is not passed', () => {
       let fn = () => {
         context
           .width(250)
-          .width("100%")
+          .width('100%')
           .width({ foo: 123 })
       };
       expect(fn).to.throw();
     });
 
-    it("height throws if a number of string is not passed", () => {
+    it('height throws if a number of string is not passed', () => {
       let fn = () => {
         context
           .height(250)
-          .height("100%")
+          .height('100%')
           .height({ foo: 123 })
       };
       expect(fn).to.throw();
@@ -97,102 +97,102 @@ describe("ThisContext", () => {
   });
 
 
-  describe("margin", function() {
-    it("it has a default value", () => {
+  describe('margin', function() {
+    it('it has a default value', () => {
       expect(_.isNumber(context.margin())).to.equal(true);
     });
 
-    it("stores values", () => {
+    it('stores values', () => {
       expect(context.margin(10).margin()).to.equal(10);
     });
 
-    it("throws if a number of string is not passed", () => {
+    it('throws if a number of string is not passed', () => {
       expect(() => { context.margin({}) }).to.throw();
     });
   });
 
 
-  describe("align", function() {
-    it("has a default value", () => {
-      expect(context.align()).to.equal("center top");
+  describe('align', function() {
+    it('has a default value', () => {
+      expect(context.align()).to.equal('center top');
     });
 
-    it("takes a value", () => {
-      expect(context.align("center middle").align()).to.equal("center middle");
+    it('takes a value', () => {
+      expect(context.align('center middle').align()).to.equal('center middle');
     });
 
-    it("throws if a string is not specified", () => {
+    it('throws if a string is not specified', () => {
       expect(() => { context.align(false) }).to.throw();
     });
 
-    it("throws if a non-supported enum value is specified", () => {
-      expect(() => { context.align("top foo") }).to.throw();
+    it('throws if a non-supported enum value is specified', () => {
+      expect(() => { context.align('top foo') }).to.throw();
     });
   });
 
 
-  describe("header", function() {
-    it("is undefined by default", () => {
+  describe('header', function() {
+    it('is undefined by default', () => {
       expect(context.header()).to.equal(undefined);
     });
 
-    it("can be set to null", () => {
+    it('can be set to null', () => {
       context.header(null);
       expect(context.header()).to.equal(null);
     });
 
-    it("throws if a string is not specified", () => {
+    it('throws if a string is not specified', () => {
       expect(() => { context.margin(false) }).to.throw();
     });
   });
 
 
-  describe("header.hr", function() {
-    it("is [undefined] by default", () => {
+  describe('header.hr', function() {
+    it('is [undefined] by default', () => {
       expect(context.hr()).to.equal(true);
     });
 
-    it("can be set to true", () => {
-      context.header("## My subtitle").hr(true);
+    it('can be set to true', () => {
+      context.header('## My subtitle').hr(true);
       expect(context.hr()).to.equal(true);
     });
 
-    it("throws if not boolean", () => {
-      expect(() => { context.hr("hello") }).to.throw();
+    it('throws if not boolean', () => {
+      expect(() => { context.hr('hello') }).to.throw();
     });
   });
 
 
-  describe("backdrop", function() {
-    it("has default value", () => {
+  describe('backdrop', function() {
+    it('has default value', () => {
       expect(context.backdrop()).to.equal(0);
     });
 
-    it("throws if not number or string", () => {
-      context.backdrop(0).backdrop("red");
+    it('throws if not number or string', () => {
+      context.backdrop(0).backdrop('red');
       let fn = () => { context.backdrop({}); };
       expect(fn).to.throw();
     });
   });
 
 
-  describe("scroll", function() {
-    it("is not scrolling by default", () => {
+  describe('scroll', function() {
+    it('is not scrolling by default', () => {
       expect(context.scroll()).to.equal(false);
     });
 
-    it("can be x, y or x:y", () => {
-      expect(context.scroll("x").scroll()).to.equal("x");
-      expect(context.scroll("y").scroll()).to.equal("y");
-      expect(context.scroll("x:y").scroll()).to.equal("x:y");
+    it('can be x, y or x:y', () => {
+      expect(context.scroll('x').scroll()).to.equal('x');
+      expect(context.scroll('y').scroll()).to.equal('y');
+      expect(context.scroll('x:y').scroll()).to.equal('x:y');
     });
 
-    it("can be true/false", () => {
+    it('can be true/false', () => {
       expect(context.scroll(false).scroll()).to.equal(false);
       expect(context.scroll(true).scroll()).to.equal(true);
     });
 
-    it("throws if not supported value", () => {
+    it('throws if not supported value', () => {
       let fn = () => { context.scroll({}) };
       expect(fn).to.throw();
     });